Add set_symlink_file_times that doesn't follow symlinks
This adds the method set_symlink_file_times that doesn't follow
symlinks and instead sets the time of the symlink itself.

Because implementations of set_file_times and set_symlink_file_times
are so similar for both windows and unix I have also refactor the code
be adding set_file_times_w for windows and set_file_times_u for unix
that contains the shared code for the two functions.